Issue #23, Display question now has tooltips. Also some work on Issue #28, added beginnings of unit test for ExamManager.. Mockito is included, but not used, as it turns out its not necessary in this test

+++ src/main/java/com/haxwell/apps/questions/managers/ExamManager.java
@@ -87,7 +87,7 @@ public class ExamManager extends Manager {
 		while (iterator.hasNext()){
 			Question q = iterator.next();
 			
-			if (questionIds.contains(q.getId()))
+			if (questionIds.contains((int)q.getId()))
 			{
 				list.add(q);
 			}
